Release v13.3.0 - Squashed schema, table prefixing, adjustments/dimensions ledgers, normalized return reasons
BREAKING CHANGE: single squashed migration (database/migrations_squashed) replaces all incremental migrations (moved to database/migrations_legacy, reference only).

- Configurable table prefix/names via config('commerce.general.prefix') + CommerceTables::name(); BaseModel::getTable() resolves through it.

- orders/invoices no longer have discount_amount/tax_amount/shipping_amount columns; document_adjustments (polymorphic) is now the single source of truth, exposed via HasAdjustments (shippingAmount()/taxAmount()/discountAmount()).

- document_contexts renamed to document_dimensions (polymorphic documentable) with HasDimensions trait; sales_unit_id/warehouse_id shortcuts plus generic addDimension().

- order_return_lines.reason replaced by normalized return_reason_id (FK to new return_reasons table) + reason_note.

- wallets.branch_id is now NOT NULL with convention 0 = global.

- New database/seeders/CommerceSeeder.php seeds default payment_methods, shipping_methods, return_reasons (publishable via commerce-seeders tag).

- Removed calculator contracts (TaxCalculatorContract/DiscountCalculatorContract/TotalsCalculatorContract) in favor of adjustments.

- Docs rewritten across usage/ and concepts/ to reflect the new schema and APIs.

vendor/bin/phpunit: 43 tests, 3220 assertions, OK. vendor/bin/pint --test: passed.

+++ src/Models/Discount.php
@@ -30,8 +30,8 @@ class Discount extends BaseModel
         return [
             'type' => DiscountTypeEnum::class,
             'value' => 'float',
-            'min_order_amount' => 'float',
-            'max_discount_amount' => 'float',
+            'min_order_amount' => 'integer',
+            'max_discount_amount' => 'integer',
             'usage_limit' => 'integer',
             'usage_per_user' => 'integer',
             'used_count' => 'integer',
